requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-apache-spark==4.1.5
apache-airflow-providers-openlineage>=1.8.0
pyspark==3.4.0
Faker
pymongo[srv]==4.6.1
qdrant-client==1.7.0
numpy>=1.21.0
scikit-learn==1.0.2
kafka-python
confluent-kafka==2.3.0
dnspython==2.4.2
minio==7.1.16
torch==2.0.1
psutil
fastembed
